requirements.txt
# Playwright is a library for browser automation, useful for testing web applications.
playwright~=1.56.0

# Ollama is a package for working with machine learning models, providing an interface for model management.
ollama~=0.6.1

# Python-dotenv is used to load environment variables from a .env file, helping manage configuration settings.
python-dotenv~=1.2.1

# Certifi provides a curated collection of root certificates for validating the trustworthiness of SSL certificates.
certifi~=2025.11.12

# Python-certifi-win32 is a Windows-specific package that provides the same functionality as certifi for Windows users.
python-certifi-win32~=1.6.1
